Fix bug for user sign up/log in
