+++ frontend/public/index.html
@@ -1,1 +1,1 @@
-<!doctype html><html lang="en"><head><meta charset="UTF-8"><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="viewport" content="width=device-width,initial-scale=1"><title>Howler</title><script defer="defer" src="/bundledd42e97de2772bfe28da.js"></script></head><body><div id="root"></div></body></html>
+<!doctype html><html lang="en"><head><meta charset="UTF-8"><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="viewport" content="width=device-width,initial-scale=1"><title>Howler</title><script defer="defer" src="/bundlecd7a7b2933cc4d7a7bfe.js"></script></head><body><div id="root"></div></body></html>
